+++ apps/expo-go/android/expoview/src/main/java/host/exp/exponent/headless/InternalHeadlessAppLoader.kt
@@ -84,7 +84,9 @@ class InternalHeadlessAppLoader(private val context: Context) :
         override fun onManifestCompleted(manifest: Manifest) {
           Exponent.instance.runOnUiThread {
             try {
-              val bundleUrl = ExponentUrls.toHttp(manifest.getBundleURL())
+              val bundleUrl = ExponentUrls.toHttp(
+                ExponentUrls.resolveManifestUrl(manifest.getBundleURL(), manifestUrl!!)
+              )
               activityIdToBundleUrl.put(activityId, bundleUrl)
               setManifest(manifestUrl!!, manifest, bundleUrl)
             } catch (e: JSONException) {